Refactor AudioIODevice to use new streaming classes.

+++ SerialPrograms/Source/CommonFramework/AudioPipeline/AudioConstants.h
@@ -12,9 +12,9 @@
 //  Largest FFT that is no more than 1/10 of a second.
 const int FFT_LENGTH_POWER_OF_TWO = 12;
 
-const int NUM_FFT_SAMPLES = 1 << FFT_LENGTH_POWER_OF_TWO;
+const size_t NUM_FFT_SAMPLES = 1 << FFT_LENGTH_POWER_OF_TWO;
 //const int NUM_FFT_WINDOWS = 100;
-const int FFT_SLIDING_WINDOW_STEP = NUM_FFT_SAMPLES/4;
+const size_t FFT_SLIDING_WINDOW_STEP = NUM_FFT_SAMPLES/4;
 
 
 #endif
